==> 0-Environment Setup.R <==
# ===== Introduction =====
# Project: Learning R by Doing - HarvardX-MITx Analysis
# By: Matthew Hendrickson
# What: Environment Setup

# ===== Install needed packages =====
install.packages("tidyverse")
install.packages("forcats")
install.packages("binr")



# ===== Load needed packages =====
library(tidyverse)
library(forcats)
library(scales)
library(lubridate)
library(binr)
